refactor(apim): refactor model to interface
