+++ packages/fx-core/src/plugins/resource/apim/index.ts
@@ -3,11 +3,11 @@
 import { Plugin, FxError, PluginContext, SystemError, UserError, Result, err, ok, QTreeNode, Stage, Func } from "@microsoft/teamsfx-api";
 import { BuildError, UnhandledError } from "./error";
 import { Telemetry } from "./utils/telemetry";
-import { AadPluginConfig, ApimPluginConfig, FunctionPluginConfig, SolutionConfig } from "./model/config";
+import { AadPluginConfig, ApimPluginConfig, FunctionPluginConfig, SolutionConfig } from "./config";
 import { AadDefaultValues, PluginLifeCycle, PluginLifeCycleToProgressStep, ProgressMessages, ProgressStep, ProjectConstants, OperationStatus } from "./constants";
 import { Factory } from "./factory";
 import { ProgressBar } from "./utils/progressBar";
-import { buildAnswer } from "./model/answer";
+import { buildAnswer } from "./answer";
 
 export class ApimPlugin implements Plugin {
     private progressBar: ProgressBar = new ProgressBar();
