+++ packages/fx-core/src/plugins/resource/apim/interfaces/IAadResource.ts
