+++ packages/fx-core/src/plugins/resource/apim/interfaces/IApimResource.ts
