+++ packages/fx-core/src/plugins/resource/apim/interfaces/IName.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+export interface IName {
+    shortName: string;
+    displayName: string;
+}
