+++ packages/fx-core/src/plugins/resource/apim/interfaces/IOpenApiDocument.ts
@@ -1,11 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
-
 import { OpenAPI } from "openapi-types";
-export enum OpenApiSchemaVersion {
-    V2 = "v2",
-    V3 = "v3",
-}
+import { OpenApiSchemaVersion } from "../constants";
 
 export interface IOpenApiDocument {
     schemaVersion: OpenApiSchemaVersion;
