+++ packages/fx-core/src/plugins/resource/apim/managers/apimManager.ts
@@ -2,10 +2,10 @@
 // Licensed under the MIT license.
 import { ApimDefaultValues, ApimPluginConfigKeys, TeamsToolkitComponent } from "../constants";
 import { AssertConfigNotEmpty, AssertNotEmpty } from "../error";
-import { IAadPluginConfig, IApimPluginConfig, IFunctionPluginConfig, ISolutionConfig } from "../model/config";
-import { ApimService } from "../service/apimService";
+import { IAadPluginConfig, IApimPluginConfig, IFunctionPluginConfig, ISolutionConfig } from "../config";
+import { ApimService } from "../services/apimService";
 import { OpenApiProcessor } from "../utils/openApiProcessor";
-import { IAnswer } from "../model/answer";
+import { IAnswer } from "../answer";
 import { LogProvider, TelemetryReporter } from "@microsoft/teamsfx-api";
 import { Lazy } from "../utils/commonUtils";
 import { NamingRules } from "../utils/namingRules";
