+++ packages/fx-core/src/plugins/resource/apim/managers/teamsAppAadManager.ts
@@ -3,8 +3,8 @@
 import { LogProvider, TelemetryReporter } from "@microsoft/teamsfx-api";
 import { ApimPluginConfigKeys, TeamsToolkitComponent } from "../constants";
 import { AssertConfigNotEmpty } from "../error";
-import { IApimPluginConfig, IAadPluginConfig } from "../model/config";
-import { AadService } from "../service/aadService";
+import { IApimPluginConfig, IAadPluginConfig } from "../config";
+import { AadService } from "../services/aadService";
 import { Lazy } from "../utils/commonUtils";
 
 export class TeamsAppAadManager {
