--- packages/fx-core/src/plugins/resource/apim/model/errorHandlerResult.ts
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-export enum ErrorHandlerResult {
-    Continue = "Continue",
-    Return = "Return",
-}
