+++ packages/fx-core/tests/plugins/resource/apim/aadManager.test.ts
@@ -6,13 +6,13 @@ import chai from "chai";
 import chaiAsPromised from "chai-as-promised";
 import sinon from "sinon";
 import dotenv from "dotenv";
-import { AadManager } from "../../../../src/plugins/resource/apim/manager/aadManager";
+import { AadManager } from "../../../../src/plugins/resource/apim/managers/aadManager";
 import { v4 } from "uuid";
 import { AadHelper, MockGraphTokenProvider, it_if, after_if, before_if, EnvConfig } from "./testUtil";
 import { InvalidAadObjectId } from "../../../../src/plugins/resource/apim/error";
-import { IRequiredResourceAccess } from "../../../../src/plugins/resource/apim/model/aadResource";
-import { AadService } from "../../../../src/plugins/resource/apim/service/aadService";
-import { IAadPluginConfig, IApimPluginConfig } from "../../../../src/plugins/resource/apim/model/config";
+import { IRequiredResourceAccess } from "../../../../src/plugins/resource/apim/interfaces/IAadResource";
+import { AadService } from "../../../../src/plugins/resource/apim/services/aadService";
+import { IAadPluginConfig, IApimPluginConfig } from "../../../../src/plugins/resource/apim/config";
 import { AadDefaultValues } from "../../../../src/plugins/resource/apim/constants";
 import { Lazy } from "../../../../src/plugins/resource/apim/utils/commonUtils";
 dotenv.config();
