+++ packages/fx-core/tests/plugins/resource/apim/aadService.test.ts
@@ -7,9 +7,9 @@ import chaiAsPromised from "chai-as-promised";
 import sinon from "sinon";
 import { v4 } from "uuid";
 import { AadHelper, MockGraphTokenProvider, it_if, before_if, after_if, EnvConfig } from "./testUtil";
-import { AadService } from "../../../../src/plugins/resource/apim/service/aadService";
+import { AadService } from "../../../../src/plugins/resource/apim/services/aadService";
 import { AadDefaultValues } from "../../../../src/plugins/resource/apim/constants";
-import { IAadInfo } from "../../../../src/plugins/resource/apim/model/aadResource";
+import { IAadInfo } from "../../../../src/plugins/resource/apim/interfaces/IAadResource";
 chai.use(chaiAsPromised);
 
 const UT_SUFFIX = v4().substring(0, 6);
