+++ packages/fx-core/tests/plugins/resource/apim/config.test.ts
@@ -3,7 +3,7 @@
 import "mocha";
 import chai from "chai";
 import { ApimPluginConfigKeys, TeamsToolkitComponent, SolutionConfigKeys } from "../../../../src/plugins/resource/apim/constants";
-import { ApimPluginConfig, SolutionConfig } from "../../../../src/plugins/resource/apim/model/config";
+import { ApimPluginConfig, SolutionConfig } from "../../../../src/plugins/resource/apim/config";
 import { ConfigValue, PluginIdentity, ReadonlyPluginConfig } from "@microsoft/teamsfx-api";
 
 describe("config", () => {
