+++ packages/fx-core/tests/plugins/resource/apim/openApiProcessor.test.ts
@@ -5,8 +5,8 @@ import chai from "chai";
 import chaiAsPromised from "chai-as-promised";
 import path from "path";
 import { OpenApiProcessor } from "../../../../src/plugins/resource/apim/utils/openApiProcessor";
-import { OpenApiSchemaVersion } from "../../../../src/plugins/resource/apim/model/openApiDocument";
 import { InvalidFunctionEndpoint, InvalidOpenApiDocument } from "../../../../src/plugins/resource/apim/error";
+import { OpenApiSchemaVersion } from "../../../../src/plugins/resource/apim/constants";
 chai.use(chaiAsPromised);
 
 const testDataBaseFolder = "./tests/plugins/resource/apim/data/openApiProcessor";
