+++ packages/fx-core/tests/plugins/resource/apim/teamsAppAadManager.test.ts
@@ -6,9 +6,9 @@ import chaiAsPromised from "chai-as-promised";
 import sinon from "sinon";
 import dotenv from "dotenv";
 import { MockGraphTokenProvider, it_if, before_if, after_if, AadHelper, EnvConfig } from "./testUtil";
-import { AadService } from "../../../../src/plugins/resource/apim/service/aadService";
-import { IAadPluginConfig } from "../../../../src/plugins/resource/apim/model/config";
-import { TeamsAppAadManager } from "../../../../src/plugins/resource/apim/manager/teamsAppAadManager";
+import { AadService } from "../../../../src/plugins/resource/apim/services/aadService";
+import { IAadPluginConfig } from "../../../../src/plugins/resource/apim/config";
+import { TeamsAppAadManager } from "../../../../src/plugins/resource/apim/managers/teamsAppAadManager";
 import axios, { AxiosInstance } from "axios";
 import { AadDefaultValues } from "../../../../src/plugins/resource/apim/constants";
 import { assert } from "sinon";
